+++ misskey-api/src/endpoint.rs
@@ -117,3 +117,7 @@ pub mod roles;
 #[cfg(feature = "13-10-0")]
 #[cfg_attr(docsrs, doc(cfg(feature = "13-10-0")))]
 pub mod renote_mute;
+
+#[cfg(feature = "13-10-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-10-0")))]
+pub mod emoji;