+++ misskey-api/src/endpoint/antennas/delete.rs
@@ -25,6 +25,7 @@ mod tests {
             .test(
                 crate::endpoint::antennas::create::Request::builder()
                     .name("test")
+                    .keywords("test")
                     .build(),
             )
             .await;