+++ misskey-api/src/endpoint/notes/create.rs
@@ -1,5 +1,7 @@
 #[cfg(feature = "12-47-0")]
 use crate::model::channel::Channel;
+#[cfg(feature = "13-10-0")]
+use crate::model::note::ReactionAcceptance;
 use crate::model::{
     drive::DriveFile,
     id::Id,
@@ -69,6 +71,11 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub local_only: Option<bool>,
+    #[cfg(feature = "13-10-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "13-10-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub reaction_acceptance: Option<ReactionAcceptance>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub no_extract_mentions: Option<bool>,
@@ -125,6 +132,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -150,6 +159,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: Some(true),
                 local_only: Some(true),
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: Some(true),
                 no_extract_hashtags: Some(true),
                 no_extract_emojis: Some(true),
@@ -175,6 +186,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -202,6 +215,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -222,6 +237,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -242,6 +259,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -265,6 +284,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -273,6 +294,51 @@ mod tests {
                 renote_id: None,
                 poll: None,
                 #[cfg(feature = "12-47-0")]
+                channel_id: None,
+            })
+            .await;
+    }
+
+    #[cfg(feature = "13-10-0")]
+    #[tokio::test]
+    async fn request_with_reaction_acceptance() {
+        use crate::model::note::ReactionAcceptance;
+
+        let client = TestClient::new();
+        client
+            .test(Request {
+                visibility: None,
+                visible_user_ids: None,
+                text: Some("nobody can react me".to_string()),
+                cw: None,
+                local_only: None,
+                reaction_acceptance: Some(ReactionAcceptance::LikeOnly),
+                no_extract_mentions: None,
+                no_extract_hashtags: None,
+                no_extract_emojis: None,
+                file_ids: None,
+                reply_id: None,
+                renote_id: None,
+                poll: None,
+                channel_id: None,
+            })
+            .await;
+        let client = TestClient::new();
+        client
+            .test(Request {
+                visibility: None,
+                visible_user_ids: None,
+                text: Some("remote cannot react me".to_string()),
+                cw: None,
+                local_only: None,
+                reaction_acceptance: Some(ReactionAcceptance::LikeOnlyForRemote),
+                no_extract_mentions: None,
+                no_extract_hashtags: None,
+                no_extract_emojis: None,
+                file_ids: None,
+                reply_id: None,
+                renote_id: None,
+                poll: None,
                 channel_id: None,
             })
             .await;
@@ -290,6 +356,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -311,6 +379,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -336,6 +406,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -357,6 +429,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -397,6 +471,8 @@ mod tests {
                     #[cfg(not(feature = "12-96-0"))]
                     via_mobile: None,
                     local_only: None,
+                    #[cfg(feature = "13-10-0")]
+                    reaction_acceptance: None,
                     no_extract_mentions: None,
                     no_extract_hashtags: None,
                     no_extract_emojis: None,
@@ -426,6 +502,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,
@@ -460,6 +538,8 @@ mod tests {
                 #[cfg(not(feature = "12-96-0"))]
                 via_mobile: None,
                 local_only: None,
+                #[cfg(feature = "13-10-0")]
+                reaction_acceptance: None,
                 no_extract_mentions: None,
                 no_extract_hashtags: None,
                 no_extract_emojis: None,