+++ misskey-api/src/test.rs
@@ -102,26 +102,12 @@ impl<T: Client + Send + Sync> ClientExt for T {
         renote_id: Option<Id<Note>>,
         reply_id: Option<Id<Note>>,
     ) -> Note {
-        self.test(crate::endpoint::notes::create::Request {
-            visibility: None,
-            visible_user_ids: None,
-            text: text.map(|x| x.to_string()),
-            cw: None,
-            #[cfg(not(feature = "12-96-0"))]
-            via_mobile: None,
-            local_only: None,
-            no_extract_mentions: None,
-            no_extract_hashtags: None,
-            no_extract_emojis: None,
-            file_ids: None,
-            reply_id,
-            renote_id,
-            poll: None,
-            #[cfg(feature = "12-47-0")]
-            channel_id: None,
-        })
-        .await
-        .created_note
+        let mut request = crate::endpoint::notes::create::Request::builder().build();
+        request.text = text.map(Into::into);
+        request.renote_id = renote_id;
+        request.reply_id = reply_id;
+
+        self.test(request).await.created_note
     }
 
     async fn avatar_url(&self) -> Url {
@@ -158,13 +144,12 @@ impl<T: Client + Send + Sync> ClientExt for T {
 
         loop {
             let files = self
-                .test(crate::endpoint::drive::files::Request {
-                    type_: None,
-                    folder_id: Some(folder.id),
-                    limit: Some(1),
-                    since_id: None,
-                    until_id: None,
-                })
+                .test(
+                    crate::endpoint::drive::files::Request::builder()
+                        .folder_id(folder.id)
+                        .limit(1)
+                        .build(),
+                )
                 .await;
             if let Some(file) = files.into_iter().next() {
                 break file;